experiment/agentar_scale_sql/requirements.txt
chromadb==1.0.15
func_timeout==4.3.5
langchain==0.3.27
langchain_core==0.3.72
langchain_openai==0.3.28
langfuse==3.2.1
langgraph==0.6.1
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
sentence_transformers==5.0.0
tiktoken==0.9.0
tabulate==0.9.0
modelscope==1.28.2
pyserini==1.2.0
ijson==3.4.0
nltk==3.9.1
sqlglot==27.13.2